backend/pyproject.toml
[project]
name = "childsafe-rag"
version = "0.1.0"
dependencies = [
    "click>=8.3.0",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "langgraph-supervisor>=0.0.29",
    "langsmith>=0.4.29",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "fastapi",
    "uvicorn",
    "chromadb",
    "pypdf",
    "pytest",
    "requests>=2.32.5",
    "pypdf2>=3.0.1",
    "fabric>=3.2.2",
    "langchain>=0.3.27",
]
requires-python = ">=3.10"
